Add search logic to GetAllStudentQuery

+++ StudentsAPI/Services/University/Queries/GetAllUniversitiesQuery.cs
@@ -22,7 +22,7 @@ namespace StudentsAPI.Services.University.Queries
                 CancellationToken cancellationToken)
             {
                 var universitiesList = await _context.Universities
-                    .Skip(request.Pagination.ItemsPerPage * request.Pagination.Page)
+                    .Skip((int) (request.Pagination.ItemsPerPage * request.Pagination.Page))
                     .Take(request.Pagination.ItemsPerPage)
                     .ToListAsync(cancellationToken: cancellationToken);
